requirements.txt
absl-py==1.3.0
addict==2.4.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==20.2.0
autopep8==2.0.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorlover==0.3.0
contourpy==1.0.6
cycler==0.11.0
dash==2.7.1
dash-bootstrap-components==1.2.1
dash-colorscales==0.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.6
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.9.1
fastjsonschema==2.16.1
ffmpeg==1.4
ffprobe==0.5
filelock==3.8.0
Flask==2.2.2
Flask-Compress==1.13
Flask-Cors==3.0.10
flask-ngrok==0.0.25
flatbuffers==22.10.26
fonttools==4.34.4
future==0.18.2
gast==0.4.0
gdown==4.5.1
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.50.0
h5py==3.7.0
idna==3.3
imageio==2.22.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.0.1
ipykernel==6.16.0
ipympl==0.9.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.14.0
jupyter-client==7.4.2
jupyter-core==4.11.1
jupyter-packaging==0.12.3
jupyter-server==1.21.0
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.2
jupyterlab-widgets==1.1.1
kaleido==0.2.1
kiwisolver==1.4.4
kornia==0.6.7
libclang==14.0.6
llvmlite==0.39.1
lxml==4.9.1
lz4==4.0.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.3
missingno==0.4.2
mistune==2.0.4
more-itertools==8.5.0
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.8
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.4
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
panda3d-gltf==0.13
panda3d-simplepbr==0.10
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.7.post1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
plotly==5.10.0
pluggy==0.13.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycolmap==0.3.0
pycparser==2.21
pydub==0.25.1
Pygments==2.12.0
pyngrok==5.1.0
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.0.2
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qrcode==7.3.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.12.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
squarify==0.4.3
stack-data==0.3.0
statsmodels==0.13.5
strsimpy==0.2.1
tabulate==0.9.0
tenacity==8.0.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.1.1
toml==0.10.1
tomli==2.0.1
tomlkit==0.11.4
torch==1.13.0
torch-tb-profiler==0.4.0
torchaudio==0.13.0
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing-extensions==4.3.0
urllib3==1.26.12
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wrapt==1.14.1
xarray==2022.12.0
xlrd==2.0.1
zipp==3.1.0

